ActionConfigureComponents moved to actions folder
