+++ ui-tests/src/test/java/io/syndesis/qe/pages/SyndesisPage.java
@@ -8,11 +8,11 @@ import io.syndesis.qe.pages.customizations.CustomizationsPage;
 import io.syndesis.qe.pages.customizations.extensions.TechExtensionDetailPage;
 import io.syndesis.qe.pages.customizations.extensions.TechExtensionsImportPage;
 import io.syndesis.qe.pages.customizations.extensions.TechExtensionsListComponent;
-import io.syndesis.qe.pages.integrations.edit.ActionConfigureComponentInvokeSql;
-import io.syndesis.qe.pages.integrations.edit.ActionConfigureComponentJmsPublish;
-import io.syndesis.qe.pages.integrations.edit.ActionConfigureComponentJmsRequest;
-import io.syndesis.qe.pages.integrations.edit.ActionConfigureComponentPeriodicSql;
-import io.syndesis.qe.pages.integrations.edit.ActionConfigureComponentJmsSubscribe;
+import io.syndesis.qe.pages.integrations.edit.actions.ActionConfigureComponentInvokeSql;
+import io.syndesis.qe.pages.integrations.edit.actions.ActionConfigureComponentJmsPublish;
+import io.syndesis.qe.pages.integrations.edit.actions.ActionConfigureComponentJmsRequest;
+import io.syndesis.qe.pages.integrations.edit.actions.ActionConfigureComponentPeriodicSql;
+import io.syndesis.qe.pages.integrations.edit.actions.ActionConfigureComponentJmsSubscribe;
 import io.syndesis.qe.pages.integrations.edit.IntegrationConnectionSelectComponentFinish;
 import io.syndesis.qe.pages.integrations.edit.IntegrationConnectionSelectComponentStart;
 import io.syndesis.qe.pages.integrations.edit.IntegrationSaveOrAddStepComponent;
