+++ ui-tests/src/test/java/io/syndesis/qe/pages/integrations/edit/actions/ActionConfigureComponentInvokeSql.java
@@ -1,4 +1,4 @@
-package io.syndesis.qe.pages.integrations.edit;
+package io.syndesis.qe.pages.integrations.edit.actions;
 
 import static com.codeborne.selenide.Condition.visible;
 import static com.codeborne.selenide.Selenide.$;
