+++ ui-tests/src/test/java/io/syndesis/qe/pages/integrations/edit/actions/ActionConfigureComponentJmsPublish.java
@@ -1,4 +1,4 @@
-package io.syndesis.qe.pages.integrations.edit;
+package io.syndesis.qe.pages.integrations.edit.actions;
 
 import static com.codeborne.selenide.Condition.visible;
 
@@ -10,22 +10,21 @@ import lombok.extern.slf4j.Slf4j;
  * Created by sveres on 12/19/17.
  */
 @Slf4j
-public class ActionConfigureComponentJmsSubscribe extends ActionConfigureComponentJms {
-
+public class ActionConfigureComponentJmsPublish extends ActionConfigureComponentJms {
 
 	private static final class Element {
-		public static final By TITLE = By.cssSelector("div[innertext='Subscribe for Messages']");
+		public static final By TITLE = By.cssSelector("div[innertext='Publish Messages']");
 	}
 
 	private static final class Input {
 		public static final By DESTINATION_NAME = By.id("destinationName");
-		public static final By SUBCRIPTION_ID = By.id("durableSubscriptionId");
-		public static final By MESSAGE_SELECTOR = By.id("messageSelector");
+		public static final By PERSISTENT = By.id("persistent");
 	}
 
 	private static final class Select {
 		public static final By DESTINATION_TYPE = By.id("destinationType");
 	}
+
 
 	@Override
 	public boolean validate() {
@@ -34,11 +33,11 @@ public class ActionConfigureComponentJmsSubscribe extends ActionConfigureCompone
 
 	@Override
 	public Class getInputClass() {
-		return ActionConfigureComponentJmsSubscribe.Input.class;
+		return ActionConfigureComponentJmsPublish.Input.class;
 	}
 
 	@Override
 	public Class getSelectClass() {
-		return ActionConfigureComponentJmsSubscribe.Select.class;
+		return ActionConfigureComponentJmsPublish.Select.class;
 	}
 }
